add parsing stuff

+++ src/lua/invader2kaitai.lua
@@ -7,7 +7,9 @@ local convert = require "convert"
 
 local parser = argparse("invader2kaitai", "Convert Invader tags to Kaitai Struct")
 parser:argument("input", "Input file"):args(1)
-parser:flag("-r --runtime", "Generate runtime structure")
+-- parser:flag("-r --runtime", "Generate runtime structure")
+parser:flag("-c --code", "Generate language code")
+parser:flag("-s --struct", "Dump KSY result to a C struct")
 
 local args = parser:parse()
 
@@ -17,72 +19,17 @@ local defFileName = "tag_" .. fileName
 local defName = defFileName:match("^(.+)%..+$")
 local defLabel = fileName:match("^(.+)%..+$")
 
-local primitiveTypes = {
-    u1 = "std::uint8_t",
-    u2 = "std::uint16_t",
-    u4 = "std::uint32_t",
-    u8 = "std::uint64_t",
-    s1 = "std::int8_t",
-    s2 = "std::int16_t",
-    s4 = "std::int32_t",
-    s8 = "std::int64_t",
-    f4 = "float",
-    f8 = "double",
-    str = "char"
-}
-
-local function mapToCStruct(structName, structElements, members)
-    local struct = [[
-#include <cstdint>
-
-namespace Balltze::Engine { 
-    ]]
-    local function add(text)
-        struct = struct .. [[
-    ]] .. text
-    end
-    for _, element in pairs(structElements) do
-        add("struct " .. convert.snakeCaseToCamelCase(element.id) .. " {")
-        for _, sequenceElement in ipairs(element.seq) do
-            if sequenceElement.id then
-                local type = sequenceElement.type
-                if primitiveTypes[type] then
-                    type = primitiveTypes[type]
-                elseif type then
-                    type = convert.snakeCaseToCamelCase(type)
-                end
-                if type then
-                    add(type .. " " .. sequenceElement.id .. ";\n")
-                else
-                    add("std::byte " .. sequenceElement.id .. "[" .. sequenceElement.size .. "];\n")
-                end
-            end
-        end
-        add("};\n")
-    end
-    add("}\n")
-
-    if members then
-        for _, sequenceElement in pairs(members) do
-            local type = sequenceElement.type
-            if primitiveTypes[type] then
-                type = primitiveTypes[type]
-            elseif type then
-                type = convert.snakeCaseToCamelCase(type)
-            end
-            if type then
-                add(type .. " " .. sequenceElement.id .. ";\n")
-            end
-        end
-    end
-    glue.writefile("src/include/" .. structName .. ".hpp", struct, "t")
-end
-
 local ksy = {
     meta = {id = defName, endian = "be", imports = {"tag_header"}},
     seq = {{id = "header", type = "tag_header"}},
     enums = {},
     types = {
+        tag_id = {
+            seq = {
+                {id = "index", type = "u2"},
+                {id = "id", type = "u2"},
+            }
+        },
         tag_reflexive = {
             seq = {
                 {
@@ -94,7 +41,8 @@ local ksy = {
                     id = "offset",
                     type = "u4"
                     -- doc = "Offset to elements"
-                }
+                },
+                {size = 4}
             }
         },
         point2_d = {seq = {{id = "x", type = "f4"}, {id = "y", type = "f4"}}},
@@ -128,8 +76,8 @@ local ksy = {
                 {id = "size", type = "u4"},
                 {id = "external", type = "u4"},
                 {id = "file_offset", type = "u4"},
-                {id = "pointer", size = 8}
-                -- {id = "pointer", type = "u8"}
+                {id = "pointer", type = "u4"},
+                {size = 4}
             }
         },
         tag_dependency = {
@@ -147,9 +95,10 @@ local types = {}
 for typeName, type in pairs(ksy.types) do
     types[#types + 1] = {id = typeName, seq = type.seq}
 end
-mapToCStruct("types", types)
+--convert.toCStruct("types", types)
 
 local structs = {}
+local flags = {}
 
 local function elementToSequence(element)
     local fieldName = convert.normalToSnakeCase(convert.camelCaseToSnakeCase(element.name))
@@ -171,23 +120,17 @@ local function elementToSequence(element)
         local sequence = {type = "u2"}
     elseif element.type == "bitfield" then
         element.width = element.width or 8
-        -- if element.fields then
-        --    ksy.enums[fieldName] = {}
-        --    for index, value in pairs(element.fields) do
-        --        local value = convert.dashAndSentenceToSnakeCase(value)
-        --        if value:match("^%d") then
-        --            value = "e_" .. value
-        --        end
-        --        ksy.enums[fieldName][tostring(index - 1)] = value
-        --    end
-        -- end
         local sequence = {id = fieldName, type = "b" .. element.width, doc = element.description}
-        structs[#structs + 1] = {id = fieldName, seq = sequence}
+        -- structs[#structs + 1] = {id = fieldName, seq = sequence, fields = element.fields}
+
         ksy.types[fieldName] = {
             seq = {{id = fieldName, type = "b" .. element.width, doc = element.description}}
         }
         return sequence
     elseif element.type == "float" or element.type == "Angle" or element.type == "Fraction" then
+        if element.bounds then
+            return {id = fieldName, type = "f4", size = 2}
+        end
         return {id = fieldName, type = "f4"}
     elseif element.type == "int16" then
         return {id = fieldName, type = "s2"}
@@ -216,6 +159,8 @@ local function elementToSequence(element)
         structs[#structs + 1] = {id = fieldName, seq = sequence}
         return sequence
     elseif element.type == "TagReflexive" then
+        -- FIXME We need to remove the scrupt property from this
+        --return {id = fieldName, type = "tag_reflexive", struct = element.struct}
         return {id = fieldName, type = "tag_reflexive"}
     elseif ksy.enums[fieldType] then
         return {id = fieldName, type = "u2", doc = element.description, enum = fieldType}
@@ -229,37 +174,46 @@ local function elementToSequence(element)
     end
 end
 
-local function mapJsonProperties()
+local function tagBodyToSequence()
     for _, element in pairs(defFile) do
         print(element.name, element.type)
         local sequence = elementToSequence(element)
         if sequence and _ == #defFile then
             for _, sequenceElement in ipairs(sequence) do
-                table.insert(ksy.seq, sequenceElement)
+                local seq = glue.deepcopy(sequenceElement)
+                if seq.struct then
+                    seq.struct = nil
+                end
+                table.insert(ksy.seq, seq)
             end
         end
     end
+    ksy.enums = nil
 end
 
-mapJsonProperties()
+tagBodyToSequence()
 
-mapToCStruct(defLabel, structs, glue.map(ksy.types, function (k,v)
-    return {id = k, seq = v.seq}
-end))
+--convert.toCStruct(defLabel, structs, glue.map(ksy.types, function(k, v)
+--    return {id = k, seq = v.seq}
+--end))
 
-local savedAsJson = glue.writefile("src/defs/" .. defFileName, json.encode(ksy))
-if savedAsJson then
-    print("- Saved as JSON")
+if args.code then
+    local savedAsJson = glue.writefile("src/defs/" .. defFileName, json.encode(ksy))
+    if savedAsJson then
+        print("- Saved as JSON")
+    end
+    local convertedToYml = os.execute("cat src/defs/" .. defFileName .. " | yq -P > src/defs/" ..
+                                          defName .. ".ksy")
+    if convertedToYml then
+        print("- JSON to KSY")
+        os.remove("src/defs/" .. defFileName)
+    end
+
+    local kaitaiToLua =
+        os.execute("ksc -t lua --outdir src/lua/tags src/defs/" .. defName .. ".ksy")
+    if kaitaiToLua then
+        print("- Converted to Lua")
+    end
 end
-local convertedToYml = os.execute("cat src/defs/" .. defFileName .. " | yq -P > src/defs/" ..
-                                      defName .. ".ksy")
-if convertedToYml then
-    print("- JSON to KSY")
-    os.remove("src/defs/" .. defFileName)
-end
---local kaitaiToLua = os.execute("ksc -t lua --outdir src/lua/tags src/defs/" .. defName .. ".ksy")
---if kaitaiToLua then
---    print("- Converted to Lua")
---end
 
 print("Done")
